Moved SW360SimpleHalResource to resource sub package.
Issue: eclipse#423

This package fits better. The class is located besides other resource
base classes.
