+++ modules/sw360/sw360-client/src/main/java/org/eclipse/sw360/antenna/sw360/rest/resource/licenses/SW360License.java
@@ -12,7 +12,7 @@
 package org.eclipse.sw360.antenna.sw360.rest.resource.licenses;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import org.eclipse.sw360.antenna.sw360.rest.SW360SimpleHalResource;
+import org.eclipse.sw360.antenna.sw360.rest.resource.SW360SimpleHalResource;
 
 import java.util.Objects;
